Small adjustments
Most of the remaining work is in playing with the probability functions
for joining and leaving.

Firstly, there is a sudden spike in joins right in the first tilmestep,
which I need to correct.

Secondly, the people joining and the people leaving are always fairly
close to each other, which is not the kind of behaviour I want. I want
there to be a bias on joining/leaving depending on the market’s current
circumstances. Need to adjust probability functions to encourage
joining during good times, and encourage leaving during bad times. But
I also need to try and make things move a bit more slowly. It will take
a bit of time