Model is finished
Changed the way a-values are calculated when looking at social
connections. Instead of comparing connection strength to the node with
the largest number of connections, it’s instead compared to the average
number of connections in the social sphere.

If the connection strength is significantly larger than the average
number of connections (suggesting a hub, since it’s a scale-free
network), then the probability of following that node’s actions is very
high.

This results in exactly the behaviour that I want; hub nodes tend to
dictate the behaviour of the system, and there are thus large swings
and consistent collapses. Conversely, when I disable node-examination
(i.e. each node decides whether to join or leave the market based only
on the market, and not their social links), the behaviour of the system
is much more stable.