Added Watts-Strogatz Model
Can use the WS model when generating the social spheres. Results are
extremely interesting! The system behaves fairly rationally when the
network is generated using WS model (compared to irrational behaviour
in BA model), and is much more stable. Behaviour is overall fairly
similar to the behaviour seen when I simply exclude consideration of
the social sphere in BA model.

This suggests that the irrationality seen in the BA model results are
not a result of irrational agent behaviour, but rather the
irrationality is itself an emergent property of the system!

Need to just neaten the interface up a bit, to make it easier to
specify things like parameters and desired models at runtime (perhaps a
config text file), and then I can call the model complete, and just
finish off the report.